* Cosmetics * win32 tweaks to build
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296921 13f79535-47bb-0310-9956-ffa450edef68

+++ jk/native2/include/jk_service.h
@@ -195,11 +195,11 @@ struct jk_ws_service {
     char    *server_name;   
     unsigned server_port;   
     char    *server_software;
-    unsigned content_length;    /* integer that represents the content  */
+    long    content_length;     /* long that represents the content  */
                                 /* length should be 0 if unknown.        */
     unsigned is_chunked;        /* 1 if content length is unknown (chunked rq) */
     unsigned no_more_chunks;    /* 1 if last chunk has been read */
-    unsigned content_read;      /* number of bytes read */
+    long     content_read;      /* number of bytes read */
 
     /*
      * SSL information
